+++ src/python/deepgnn/graph_engine/snark/tests/temporal_test.py
@@ -6,7 +6,6 @@ import os
 import sys
 import tempfile
 from pathlib import Path
-import platform
 
 import numpy as np
 import numpy.testing as npt
@@ -17,7 +16,6 @@ from deepgnn.graph_engine.snark.decoders import JsonDecoder
 import deepgnn.graph_engine.snark.convert as convert
 import deepgnn.graph_engine.snark.server as server
 import deepgnn.graph_engine.snark.dispatcher as dispatcher
-import deepgnn.graph_engine.snark._lib as lib
 from util_test import find_free_port
 
 
@@ -98,19 +96,6 @@ def triangle_graph_data():
     data_name = triangle_graph_json(workdir.name)
     yield data_name
     workdir.cleanup()
-
-
-def get_lib_name():
-    lib_name = "libwrapper.so"
-    if platform.system() == "Windows":
-        lib_name = "wrapper.dll"
-    elif platform.system() == "Darwin":
-        lib_name = "libwrapper.dylib"
-    return os.path.join(os.path.dirname(__file__), "..", lib_name)
-
-
-def setup_module():
-    lib._LIB_PATH = get_lib_name()
 
 
 # We'll use this class for deterministic partitioning