+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `return_edge_created_ts` argument to neighbor sampling methods to return timestamps when edges connecting nodes were created.
+- `MOOC` temporal dataset.
 
 - Add PyG remote backend example.
 
 ### Fixed
 - Uniform sampling works in temporal graphs.
+- ADL path parsing to download graph data.
 
 ## [0.1.60] - 2022-04-18